[GROW-262] update package
