+++ AppPackage/Sources/MakePromise/SubViews/SetNameAndPlaceView.swift
@@ -7,6 +7,7 @@
 //
 
 import ComposableArchitecture
+import DesignSystem
 import SwiftUI
 
 typealias NameAndPlaceView = SetNameAndPlaceView
