+++ Planz/Promise/Color+Ex.swift
@@ -1,36 +1,1 @@
-//
-//  Color+Ex.swift
-//  Planz
-//
-//  Created by 한상준 on 2022/10/08.
-//  Copyright © 2022 Team-Planz. All rights reserved.
-//
 
-import SwiftUI
-
-extension Color {
-    init(hex: String) {
-        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int: UInt64 = 0
-        Scanner(string: hex).scanHexInt64(&int)
-        let addressable, red, green, blue: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (addressable, red, green, blue) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (addressable, red, green, blue) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (addressable, red, green, blue) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (addressable, red, green, blue) = (1, 1, 1, 0)
-        }
-
-        self.init(
-            .sRGB,
-            red: Double(red) / 255,
-            green: Double(green) / 255,
-            blue: Double(blue) / 255,
-            opacity: Double(addressable) / 255
-        )
-    }
-}
